fix: change price type

+++ Assets/Scripts/Claim.cs
@@ -6,6 +6,6 @@ namespace AntiqueShop
     public class Claim
     {
         [field: SerializeField] public string CustomerText { get; private set; }
-        [field: SerializeField] public int AskingPrice { get; private set; }
+        [field: SerializeField] public float AskingPrice { get; private set; }
     }
 }
